pyproject.toml
[project]
name = "clash"
version = "0.1.0"
description = "Clash Compiler"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "automata-lib>=9.1.2",
    "pyfiglet>=1.0.4",
    "pytest>=8.4.2",
]
